fix: start/end time format on Epoch page
